+++ sponge/src/core/engine.h
@@ -13,7 +13,7 @@ class Engine {
     virtual void shutdown() = 0;
 
     virtual bool onUserCreate() = 0;
-    virtual bool onUserUpdate(uint32_t elapsedTime) = 0;
+    virtual bool onUserUpdate(double elapsedTime) = 0;
     virtual bool onUserDestroy() = 0;
 
     virtual void onEvent(event::Event& event) = 0;
